# backend/app/Http/Requests/Canteen/RegisterFeedbackRequest.php
<?php

namespace App\Http\Requests\Canteen;

use App\Http\Requests\BaseRequest;

class RegisterFeedbackRequest extends BaseRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            "food_id" => ["required", "exists:food,id"],
            "comment" => ["required", "string"],
        ];
    }
}