# backend/app/Http/Requests/Canteen/UpdateFeedbackRequest.php
<?php

namespace App\Http\Requests\Canteen;

use App\Http\Requests\BaseRequest;

class UpdateFeedbackRequest extends BaseRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            "comment" => ["required", "string"],
        ];
    }
}