# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tummy Time API",
        "description": "API documentation for the Tummy Time project",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/feedbacks/{food_id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Get feedback for a specific food item",
                "description": "Retrieves all feedback entries for a given food ID.",
                "operationId": "getFeedbackByFoodId",
                "parameters": [
                    {
                        "name": "food_id",
                        "in": "path",
                        "description": "ID of the food item to get feedback for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of feedback entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Feedback"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Food not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feedbacks": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Create a new feedback entry",
                "description": "Creates a new feedback entry for a food item.",
                "operationId": "createFeedback",
                "requestBody": {
                    "description": "Feedback details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterFeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Feedback created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feedback"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feedbacks/{feedback}": {
            "delete": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Delete an existing feedback entry",
                "description": "Deletes an existing feedback entry by its ID. Only the owner or a SuperAdmin can delete the feedback.",
                "operationId": "deleteFeedback",
                "parameters": [
                    {
                        "name": "feedback",
                        "in": "path",
                        "description": "ID of the feedback to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Feedback deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feedback not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Update an existing feedback entry",
                "description": "Updates an existing feedback entry by its ID. Only the owner can update their feedback.",
                "operationId": "updateFeedback",
                "parameters": [
                    {
                        "name": "feedback",
                        "in": "path",
                        "description": "ID of the feedback to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated feedback comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feedback"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feedback not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format for generic errors (e.g., 401, 403, 404, 500)",
                "properties": {
                    "errors": {
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "errors": "Something went wrong."
                }
            },
...
            "Feedback": {
                "title": "Feedback",
                "description": "Feedback model",
                "properties": {
                    "id": {
                        "description": "Feedback ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID of the user who provided feedback",
                        "type": "integer",
                        "format": "int64"
                    },
                    "food_id": {
                        "description": "ID of the associated food item",
                        "type": "integer",
                        "format": "int64"
                    },
                    "comment": {
                        "description": "The feedback comment",
                        "type": "string"
                    },
                    "feedback_date": {
                        "description": "Timestamp of the feedback",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "user_id": 10,
                    "food_id": 20,
                    "comment": "This food was excellent, highly recommend!",
                    "feedback_date": "2023-01-01T14:30:00.000000Z"
                }
            },
...
            "RegisterFeedbackRequest": {
                "title": "RegisterFeedbackRequest",
                "description": "Request body for creating a new feedback entry",
                "required": [
                    "food_id",
                    "comment"
                ],
                "properties": {
                    "food_id": {
                        "description": "ID of the food item being commented on",
                        "type": "integer"
                    },
                    "comment": {
                        "description": "The feedback comment",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "food_id": 1,
                    "comment": "This food was excellent, highly recommend!"
                }
            },
            "UpdateFeedbackRequest": {
                "title": "UpdateFeedbackRequest",
                "description": "Request body for updating a feedback entry",
                "required": [
                    "comment"
                ],
                "properties": {
                    "comment": {
                        "description": "The updated feedback comment",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "comment": "The updated comment."
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}